Implement playlist shuffle.
The shuffle is maintained as a permutation on the playlist, that means
that when shuffle is on all songs are played exactly once only the order
is randomized.

Insertion into playlist is O(1) unfortunately song removal from the
playlist is now O(n) since we need to maintain the permutation as well.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>